requirements_dev.txt
-r requirements.txt

# Unit Test modules
pytest
pytest-cov
pytest-mock
coverage

# Technical Analysis (needed for VixHelper and other TA components)
pandas-ta-classic

# Linting modules
ruff
isort

# Security checkers
bandit
safety

# Secret Storage
python-dotenv

# Documentation (llms.txt generation for AI assistants)
sphinx-llms-txt
